# Heavily optimise moving code