Remove the use of deepcopy(), which is a very expensive operation in Python. Replace it with simply rebuilding the list of shapeBlocks manually. Improve performance of function by 10-20 times.

Example performance improvement.
- Using deepcopy() - 12,216ms (total program length - 14,643ms)
- Using new _copyBlocks() function - 489ms (total program length - 3,338ms)   

In this example, a **20x-25x performance increase!**

This was using a large object with a few hundred blocks, of different types.