@martinohanlon 2017-07-09
Who'd have thought Python's deepcopy process was so inefficient. Ill do some testing (not that I dont trust you!) and pull it in.